+++ CHANGELOG.md
@@ -10,6 +10,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed graphql queries from `BuyButton`
 - Added `OrderFormContext` to update the minicart on `BuyButton` 
 
+### Fixed
+- Top menu not being rendered as an overlay on scroll down.
+
+### Changed
+- Use tachyons classes in `AvailabilitySubscriber`.
+
+## [1.9.0] - 2018-07-26
+### Added
+- SSR to Slider component.
+
 ## [1.8.3] - 2018-07-25
 ### Fixed
 - `ProductPrice` propTypes, removed warning from console.